+++ pkg/deployments/CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Deprecations
 
 - Deprecated `20220916-batch-relayer-v4`.
+- Deprecated `20230206-weighted-pool-v3`.
 
 ## 3.2.0 (2023-02-24)